replace nextflow model dir
